requirements.txt
tqdm
numpy
termcolor
rich
lpips
pillow
torch
transformers
diffusers
telegram-send
